refactor: #388 장소 스크래핑 응답 관련 dto class name 및 package path 수정
