+++ src/main/java/com/zelusik/eatery/global/scraping/dto/KakaoPlaceScrapingInfo.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.domain.place.dto;
+package com.zelusik.eatery.global.scraping.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,7 @@ import java.util.Map;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class PlaceScrapingInfo {
+public class KakaoPlaceScrapingInfo {
 
     @Nullable
     private List<PlaceScrapingOpeningHourDto> openingHours;
@@ -21,13 +21,13 @@ public class PlaceScrapingInfo {
     @Nullable
     private String homepageUrl;
 
-    public static PlaceScrapingInfo of(List<PlaceScrapingOpeningHourDto> openingHours, String closingHours, String homepageUrl) {
-        return new PlaceScrapingInfo(openingHours, closingHours, homepageUrl);
+    public static KakaoPlaceScrapingInfo of(List<PlaceScrapingOpeningHourDto> openingHours, String closingHours, String homepageUrl) {
+        return new KakaoPlaceScrapingInfo(openingHours, closingHours, homepageUrl);
     }
 
     @SuppressWarnings("unchecked")
     // TODO: Object => List<Map> 변환 로직이 있어서 generic type casting 문제를 무시한다. 더 좋은 방법이 있다면 고려할 수 있음.
-    public static PlaceScrapingInfo from(Map<String, Object> attributes) {
+    public static KakaoPlaceScrapingInfo from(Map<String, Object> attributes) {
         List<Map<String, Object>> openingHours = (List<Map<String, Object>>) attributes.get("openingHours");
         Object closingHours = attributes.get("closingHours");
         Object homepageUrl = attributes.get("homepageUrl");
